+++ lib/flutter_gn_fancyleds_card.dart
@@ -1,5 +1,9 @@
-library;
+library flutter_gn_fancyleds_card;
 
 export 'src/controller/fancy_led_controller.dart';
-export 'src/enums/led_mode.dart';
+export 'src/controller/ambient_mode_controller.dart';
 export 'src/fancy_led_card.dart';
+export 'src/models/led_frame.dart';
+export 'src/services/ambient_mode_service.dart';
+export 'src/utils/hardware_health_utils.dart';
+export 'src/utils/image_processor_utils.dart';