+++ lib/src/enums/led_mode.dart
@@ -1,9 +1,0 @@
-enum LedMode {
-  staticColor,
-  breathing,
-  rainbow,
-  rainbowWave,
-  pulse,
-  chase,
-  ambient,
-}